struct redirected to model.go
