+++ cmd/jaeger/internal/configdocs/model.go
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 package configdocs
 
+import "encoding/json"
+
 // FieldDoc represents documentation for a struct field.
 type FieldDoc struct {
 	Name           string      `json:"-"`
@@ -22,4 +24,15 @@ type StructDoc struct {
 	Description string              `json:"description,omitempty"`
 	Properties  map[string]FieldDoc `json:"properties"`
 	Required    []string            `json:"required,omitempty"`
+}
+
+type JSONSchema struct {
+	Schema      string                     `json:"$schema"`
+	Title       string                     `json:"title"`
+	Description string                     `json:"description"`
+	Type        string                     `json:"type"`
+	Definitions map[string]interface{}     `json:"definitions"`
+	Properties  map[string]interface{}     `json:"properties"`
+	Required    []string                   `json:"required,omitempty"`
+	Extensions  map[string]json.RawMessage `json:"-"`
 }
